Add jsconfig.json
This file is needed in order for absolute imports to work, e.g. instead of

import Component from '../../../../components/Componenet';

we will be able to

import Component from 'components/Component';

See https://create-react-app.dev/docs/importing-a-component/#absolute-imports
for more info.

Note that once the migration to TypeScript is complete and there is a
tsconfig.json in the project this file should be removed.